Add SUPPORTED_SIGNALS constant on SystemExt
