+++ src/unknown/process.rs
@@ -19,10 +19,6 @@ pub struct Process {
 }
 
 impl ProcessExt for Process {
-    fn kill(&self) -> bool {
-        false
-    }
-
     fn kill_with(&self, _signal: Signal) -> Option<bool> {
         None
     }
